Revert "Convert the return values to use uint64 instead"
This reverts commit b17ecbd3de03e0698ed97cf3a2f5ab2fb44fd7ee.

+++ pkg/minikube/machine/info.go
@@ -97,8 +97,8 @@ func RemoteHostInfo(r command.Runner) (*HostInfo, error, error, error) {
 
 	var info HostInfo
 	info.CPUs = ncpus
-	info.Memory = int64(memory)
-	info.DiskSize = int64(disksize)
+	info.Memory = memory
+	info.DiskSize = disksize
 	return &info, cpuErr, memErr, diskErr
 }
